% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_gapstats}
\alias{add_gapstats}
\title{Add gap statistics to a DataFrame containing CIGAR strings}
\usage{
add_gapstats(df, cigar_column)
}
\arguments{
\item{df}{(DataFrame) containing a column with CIGAR strings}

\item{cigar_column}{(character) name of the column containing the CIGAR strings}
}
\value{
(DataFrame) with the added gap statistics
}
\description{
Add gap statistics to a DataFrame containing CIGAR strings
}
\keyword{internal}
