% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{align}
\alias{align}
\title{Align two sequences using WFA2 and return the pairwise distance}
\usage{
align(
  a,
  b,
  match = 0L,
  mismatch = 1L,
  gap = 1L,
  extend = 0L,
  gap2 = 0L,
  extend2 = 0L
)
}
\arguments{
\item{a}{(\code{character} string) first string to align}

\item{b}{(\code{character} string) second string to align}

\item{match}{(\code{integer} scalar) match score; positive is a bonus.}

\item{mismatch}{(\code{integer} scalar) mismatch score; positive is a penalty.}

\item{gap}{(\code{integer} scalar) gap opening score; positive is a penalty.
Alternatively, if \code{extend}, \code{gap2}, and \code{extend2} are all 0, then this is
the penalty for all gaps.}

\item{extend}{(\code{integer} scalar) gap extension score; positive is a penalty}

\item{gap2}{(\code{integer} scalar) alternate gap opening score for two-piece
affine gap penalty; positive is penalty.  Ignored if both \code{gap2} and
\code{extend2} are 0.}

\item{extend2}{(\code{integer} scalar) alternate gap extension score for
two-piece affine gap penalty; positive is penalty. Ignored if both \code{gap2}
and \code{extend2} are 0.}
}
\description{
WFA2 allows several alignment strategies, and this function
selects the least parameterized version possible given the inputs:
\itemize{
\item Edit : \code{gap} == \code{mismatch} != 0; \code{match} == \code{extend} == \code{gap2} == \code{extend2} == 0.
\item Indel : \code{gap} != 0; \code{mismatch} == \code{match} == \code{extend} == \code{gap2} == \code{extend2} == 0.
\item GapLinear : \code{extend} == \code{gap2} == \code{extend2} == 0; other parameters do not meet requirements for "Edit" or "Indel".
\item GapAffine : \code{gap2} == \code{extend2} == 0; other parameters do not meet requirements for "Edit", "Indel", or "GapLinear".
\item GapAffine2Pieces : parameters do not meet requirements for \code{Edit}, \code{Indel}, \code{GapLinear}, or \code{GapAffine}.
}
}
