% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_thresholds.R
\name{clean_taxonomy}
\alias{clean_taxonomy}
\title{Replace "placeholder" ranks in a taxonomy with NA}
\usage{
clean_taxonomy(
  taxonomy,
  ranks = c("kingdom", "phylum", "class", "order", "family", "genus", "species")
)
}
\arguments{
\item{taxonomy}{(\code{data.frame}) the taxonomy to clean}

\item{ranks}{(\code{character} vector) the ranks to clean}
}
\value{
a cleaned taxonomy
}
\description{
Replace "placeholder" ranks in a taxonomy with NA
}
