% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{dist_config}
\alias{dist_config}
\alias{dist_wfa2}
\alias{dist_edlib}
\alias{dist_ksw2}
\alias{dist_hybrid}
\alias{dist_hamming}
\alias{dist_usearch}
\alias{dist_file}
\title{Configuration helper for pairwise distance methods}
\usage{
dist_config(
  method = c("wfa2", "edlib", "ksw2", "hybrid", "hamming", "file", "usearch"),
  ...
)

dist_wfa2(
  match = 0L,
  mismatch = 1L,
  gap_open = 0L,
  gap_extend = 1L,
  gap_open2 = gap_open,
  gap_extend2 = gap_extend
)

dist_edlib()

dist_ksw2(
  match = 0L,
  mismatch = 1L,
  gap_open = 0L,
  gap_extend = 1L,
  gap_open2 = gap_open,
  gap_extend2 = gap_extend
)

dist_hybrid(cutoff = 0.1)

dist_hamming(min_overlap = 0L, ignore_gaps = TRUE)

dist_usearch(usearch = Sys.which("usearch"), usearch_ncpu = NULL)

dist_file(filename, by_name = TRUE)
}
\arguments{
\item{method}{(\code{character} string) method to use for distance calculations}

\item{...}{passed on to variants}

\item{match}{(non-negative \code{integer} scalar) score for a match}

\item{mismatch}{(positive \code{integer} scalar) penalty for a mismatch}

\item{gap_open}{(non-negative \code{integer} scalar) penalty for gap opening}

\item{gap_extend}{(positive \code{integer} scalar) penalty for gap extension}

\item{gap_open2}{(non-negative \code{integer} scalar) alternate penalty for gap
opening}

\item{gap_extend2}{(positive \code{integer} scalar) alternate penalty for gap
extension}

\item{cutoff}{(\code{numeric} scalar) A pairwise distance to decide when sequences
are considered "close"; when the sequences are expected to be "close", then
WFA2 in edit-distance mode is used, otherwise EdLib.}

\item{min_overlap}{(\code{integer} scalar) minimum length of overlap between two
sequences (i.e., length after subtracting end gaps in either sequence). If
the required overlap is not met, the distance is returned as 1.}

\item{ignore_gaps}{(\code{logical} flag) if \code{TRUE}, columns containing an internal
gap in one sequence are ignored, i.e. not counted as a mismatch and also not
counted towards the total aligned length. If \code{FALSE}, such columns are
counted as mismatches.  End gaps in either sequence, and internal gaps shared
by both sequences, are always ignored.}

\item{usearch}{(\code{character} path) path to the USEARCH executable}

\item{usearch_ncpu}{(\code{integer} scalar or \code{NULL}) number of threads for
USEARCH. If \code{NULL}, USEARCH will autodetect.}

\item{filename}{(\code{character} string) path to the distance matrix file}

\item{by_name}{(\code{logical} flag) if \code{TRUE}, the sequences are identified in
the file by name; if \code{FALSE}, they are identified by number, starting with 0.
Identification by number may be slightly faster and reduces the necessary
file size, but is not supported by \code{optimotu()}, only by
\code{optimize_thresholds()}.}
}
\value{
an object representing the pairwise distance method
}
\description{
Several pairwise distance methods are supported, all of which are based on
pre-existing implementations which are either integrated into the package or
called as external programs.
}
\section{Functions}{
\itemize{
\item \code{dist_wfa2()}: helper function for method \code{"wfa2"}

\item \code{dist_edlib()}: helper function for method \code{"edlib"}

\item \code{dist_ksw2()}: helper function for method \code{"ksw2"}

\item \code{dist_hybrid()}: helper function for method \code{"hybrid"}

\item \code{dist_hamming()}: helper function for method \code{"hamming"}

\item \code{dist_usearch()}: helper function for method \code{"usearch"}

\item \code{dist_file()}: helper function for method \code{"file"}

}}
\section{WFA2}{
WFA2 (Marco-Sola et al., 2021) is a fast pairwise alignment method based on
the wavefront algorithm. It is the default method for similarity distances,
and is the fastest method for very similar sequences. It supports edit and
indel distances, as well as models with linear, affine, and dual-affine gap
penalties.
}

\section{Edlib}{
Edlib (Šošić & Šikić, 2017) is a fast pairwise alignment method based on the
edit distance only. It is faster than WFA2 for divergent sequences.
}

\section{Hybrid}{
The hybrid method is a combination of WFA2 and Edlib. It is used with edit
distance, dispatching to WFA2 for distances smaller than a cutoff threshold,
and to Edlib for distances greater than or equal to the cutoff threshold.
}

\section{KSW2}{
KSW2 is the library used by minimap2 (Li, 2018) for pairwise alignment,
based on the diagonal formulation of Suzuki & Kasahara (2018). It supports
linear, affine, and dual-affine gap penalties, and is faster than WFA2 for
these models for divergent sequences.
}

\section{Hamming}{
Hamming distance does not perform alignment, it simply counts the number of
mismatches. It is extremely fast, but only applicable to sequences which are
already aligned, for instance using HMM or CM alignment to a common model.
The implementation is modified from the optimized CPU version of Protax (Li
et al., 2024).
}

\section{USEARCH}{
USEARCH (Edgar, 2010) is an external program for sequence similarity search
and other amplicon sequence analysis tasks. In OptimOTU, it is used to
calculate a sparse similarity matrix. It is accelerated by using kmer
profile similarity both to filter sequence pairs and to generate an anchor
for subsequent extension alignment. Because of the kmer filtering step, it
is not guaranteed to return all sequence pairs with distances less than the
distance threshold, but it is also much faster than full alignment methods.
Through gap parameters, it can be used to calculate edit distance, as well
as linear and affine gap penalties.
}

\section{File}{
Pairwise distances can also be precomputed by any other method and loaded
from a file. The file format is a tab-separated table with the IDs of two
sequences in the first two columns, and the distance in the third column.
}

\references{
Edgar, R.C., 2010. Search and clustering orders of magnitude faster than
BLAST. Bioinformatics 26, 2460–2461.
https://doi.org/10.1093/bioinformatics/btq461

Marco-Sola, S., Moure, J.C., Moreto, M., Espinosa, A., 2021. Fast gap-affine
pairwise alignment using the wavefront algorithm. Bioinformatics 37,
456–463. https://doi.org/10.1093/bioinformatics/btaa777

Šošić, M., Šikić, M., 2017. Edlib: a C/C ++ library for fast, exact sequence
alignment using edit distance. Bioinformatics 33, 1394–1395.
https://doi.org/10.1093/bioinformatics/btw753

Suzuki, H. and Kasahara, M. (2018). Introducing difference recurrence
relations for faster semi-global alignment of long sequences. BMC
Bioinformatics, 19:45.
https://doi.org/10.1186/s12859-018-2014-8

Li, H (2018) Minimap2: pairwise alignment for nucleotide sequences.
Bioinformatics, 34:3094-3100.
https://doi.org/10.1093/bioinformatics/bty191

Li, R., Ratnasingham, S., Zarubiieva, I., Somervuo, P., & Taylor, G. W.
(2024). PROTAX-GPU: a scalable probabilistic taxonomic classification
system for DNA barcodes. Philosophical transactions of the Royal Society of
London. Series B, Biological sciences, 379(1904), 20230124.
https://doi.org/10.1098/rstb.2023.0124
}
