% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_thresholds.R
\name{find_best_threshold}
\alias{find_best_threshold}
\title{Find the optimal clustering threshold using one or more measures}
\usage{
find_best_threshold(
  k,
  c,
  threads = 1L,
  measures = c("MCC", "RI", "ARI", "FMI", "MI", "AMI", "FM")
)
}
\arguments{
\item{k}{(\code{integer} matrix) the clustering partitions at different thresholds}

\item{c}{(\code{integer} vector) the true partition}

\item{threads}{(\code{integer(1)}) the number of threads to use}

\item{measures}{(\code{character} vector) the clustering quality measures to
calculate. Supported measures are "MCC", "RI", "ARI", "FMI", "MI", "AMI", and
"FM".}
}
\value{
(\code{data.frame}) a data frame with the following columns:
\itemize{
\item \code{measure} (\code{character}) the clustering quality measure
\item \code{threshold} (\code{numeric}) the threshold which maximizes the measure
\item \code{value} (\code{numeric}) the value of the measure at the threshold
}
}
\description{
Find the optimal clustering threshold using one or more measures
}
