% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{find_usearch}
\alias{find_usearch}
\title{Try to find the usearch executable}
\usage{
find_usearch()
}
\value{
(\code{character} string) the full path to the usearch executable.
}
\description{
Try to find the usearch executable
}
