% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{intersect_length}
\alias{intersect_length}
\alias{intersect_length_string}
\title{Size of the intersection between two sorted sets}
\usage{
intersect_length(c, k)

intersect_length_string(c, k)
}
\arguments{
\item{c}{(sorted \code{integer} or \code{ character} vector) the first set to compare}

\item{k}{(sorted \code{integer} or \code{ character} vector) the second set to compare}
}
\value{
(\code{integer} count) the number of elements which occur in both sets
}
\description{
This implementation is much faster that \code{length(intersect(c, k))}. However
it assumes (without checking!) that the sets are sorted.
}
