% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_thresholds.R
\name{is_placeholder}
\alias{is_placeholder}
\title{Identify "placeholder" taxa}
\usage{
is_placeholder(taxon)
}
\arguments{
\item{taxon}{(\code{character} vector) the taxon names to check}
}
\value{
a logical vector indicating whether each taxon is a placeholder
}
\description{
Identify "placeholder" taxa
}
