% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{select_sequence}
\alias{select_sequence}
\title{Choose a subset of sequences}
\usage{
select_sequence(sequence, which, negate = FALSE, name_col = NULL)
}
\arguments{
\item{sequence}{(\code{data.frame}, \code{character}, or \code{XStringSet}) sequences to
subset}

\item{which}{(\code{integer}, \code{logical}, or \code{character}) which sequences to
select}

\item{negate}{(\code{logical}) if TRUE, select all sequences except those
specified}

\item{name_col}{(\code{character}) name of the column in a data frame which
contains the sequence names}
}
\description{
Choose a subset of sequences
}
