% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{seq_distmx_internal}
\alias{seq_distmx_internal}
\title{Search for best match(es) of query sequences in reference sequences}
\usage{
seq_distmx_internal(
  seq,
  dist_config,
  parallel_config,
  threshold,
  verbose = 0L,
  details = 0L,
  span = 0L,
  constrain = TRUE
)
}
\arguments{
\item{seq}{(named \code{character}) sequences}

\item{dist_config}{(\code{optimotu_dist_config}) distance configuration}

\item{parallel_config}{(\code{optimotu_parallel_config}) parallelization
configuration. Search methods do not treat the different methods differently;
only the number of threads is used.}

\item{threshold}{(\code{numeric}) distance threshold for the search; larger
distances are not considered}

\item{verbose}{(\code{integer}) verbosity level}

\item{details}{(\code{integer}) detail level; 0 = score only, 1 = score and gap
stats, 2 = score and cigar}

\item{span}{(\code{integer}) alignment span}
}
\value{
\code{data.frame} with columns \code{seq_id1} (first sequence ID),
\code{seq_id2} (second sequence ID), \code{score1} and \code{score2} (optimal alignment
score for the two sequences in the "prealignment" (if any) and "alignment"
stages), \code{dist1} and \code{dist2} (corresponding sequence distances). If
\code{details} is 1, also includes columns \code{align_length}, \code{n_insert},
\code{n_delete}, \code{max_insert}, and \code{max_delete}. If \code{details} is 2, also
includes column \code{cigar}.
}
\description{
Search for best match(es) of query sequences in reference sequences
}
\keyword{internal}
