% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{seq_names<-}
\alias{seq_names<-}
\title{set the names of a set of sequences}
\usage{
seq_names(seq) <- value
}
\arguments{
\item{seq}{(\code{data.frame}, \code{character}, or \code{XStringSet}) sequences to set
the names of}

\item{value}{(\code{character} vector) the names to set}
}
\value{
the sequences with the names set (invisibly)
}
\description{
set the names of a set of sequences
}
