% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{seq_names}
\alias{seq_names}
\title{get the names of a set of sequences}
\usage{
seq_names(seq, name_col = NULL)
}
\arguments{
\item{seq}{(\code{data.frame}, \code{character}, or \code{XStringSet}) sequences to get
the names of}

\item{name_col}{(\code{character}) name of the column in a data frame which
contains the sequence names}
}
\value{
(\code{character} vector) the names of the sequences
}
\description{
get the names of a set of sequences
}
