% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_search.R
\name{seq_search}
\alias{seq_search}
\title{Search for the closest match(es) to sequences in a reference database}
\usage{
seq_search(
  query,
  ref,
  threshold,
  query_id = NULL,
  ref_id = NULL,
  dist_config = dist_wfa2(),
  parallel_config = parallel_concurrent(1),
  verbose = FALSE,
  return_cigar = FALSE,
  span = c("global", "extension"),
  query_seq_idx = NULL,
  query_seq_file = NULL,
  ref_seq_idx = NULL,
  ref_seq_file = NULL,
  ...
)
}
\arguments{
\item{query}{(\code{character}, \code{data.frame}, or \code{XStringSet}) sequences to search}

\item{ref}{(\code{character}, \code{data.frame}, or \code{XStringSet}) reference sequences to
search against}

\item{threshold}{(\code{numeric} scalar) maximum distance to consider a match, in
[0, 1] where 0 is identical.}

\item{query_id}{(\code{character} vector) names for the query sequences.  If they
are already named, this will replace the names.}

\item{ref_id}{(\code{character} vector) names for the reference sequences.  If they
are already named, this will replace the names.}

\item{dist_config}{(\code{optimotu_dist_config}) configuration for calculating
distances, as returned by \code{dist_config()} or its helpers.}

\item{parallel_config}{(\code{optimotu_parallel_config}) configuration for parallel
processing, as returned by \code{parallel_config()} or its helpers.}

\item{verbose}{(\code{logical} or \code{integer} scalar) print progress messages.}

\item{return_cigar}{(\code{logical} scalar) if \code{TRUE}, return the cigar string}

\item{span}{(\code{character} string) the span of the alignment; currently
accepted values are "global" and "extension".  The default is "global".}

\item{query_seq_idx, query_seq_file}{optional subsetting for \code{query}; see
\code{\link[=seq_as_char]{seq_as_char()}}.}

\item{ref_seq_idx, ref_seq_file}{optional subsetting for \code{ref}; see
\code{\link[=seq_as_char]{seq_as_char()}}.}

\item{...}{passed to methods}
}
\value{
(\code{data.frame}) with columns "seq_id" (\code{character}),
"ref_id" (\code{character}), and "dist" (\code{numeric}) giving the distance between
the query and reference. If \code{return_cigar} is \code{TRUE}, the CIGAR string is
returned in a column "cigar" (\code{character}).
}
\description{
Search for the closest match(es) to sequences in a reference database
}
