% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{seq_search_internal}
\alias{seq_search_internal}
\title{Search for best match(es) of query sequences in reference sequences}
\usage{
seq_search_internal(
  query,
  ref,
  dist_config,
  parallel_config,
  threshold,
  verbose = 0L
)
}
\arguments{
\item{query}{(named \code{character}) query sequences}

\item{ref}{(named \code{character}) reference sequences}

\item{dist_config}{(\code{optimotu_dist_config}) distance configuration}

\item{parallel_config}{(\code{optimotu_parallel_config}) parallelization
configuration. Search methods do not treat the different methods differently;
only the number of threads is used.}

\item{threshold}{(\code{numeric}) distance threshold for the search; larger
distances are not considered}

\item{verbose}{(\code{integer}) verbosity level}
}
\value{
\code{data.frame} with columns \code{seq_id} (query sequence ID),
\code{ref_id} (reference sequence ID), and \code{dist} (distance between the query
and reference sequences)
}
\description{
Search for best match(es) of query sequences in reference sequences
}
\keyword{internal}
