% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sequence_size}
\alias{sequence_size}
\title{Get the number of sequences in a set}
\usage{
sequence_size(seq)
}
\arguments{
\item{seq}{(\code{data.frame}, \code{character}, or \code{XStringSet}) sequences to count}
}
\description{
Get the number of sequences in a set
}
