% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_linkage.R
\name{single_linkage}
\alias{single_linkage}
\title{Title}
\usage{
single_linkage(
  file,
  seqnames,
  method = c("tree", "matrix"),
  thresholds = NULL,
  dmin = NULL,
  dmax = NULL,
  dstep = NULL,
  threads = 1L,
  minsplit = 1L
)
}
\arguments{
\item{minsplit}{}
}
\description{
Title
}
