% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_thresholds.R
\name{strict_median}
\alias{strict_median}
\title{Return the median of a (presorted) vector}
\usage{
strict_median(x)
}
\arguments{
\item{x}{(\code{numeric} vector) sorted vector to calculate the median of}
}
\value{
the median of \code{x}
}
\description{
Always returns a value which is contained in \code{x}, even if \code{x} has an even
number of elements, in which case the lower of the two middle values is
returned.
}
\keyword{internal}
