% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_taxon_thresholds.R
\name{superrank_thresholds}
\alias{superrank_thresholds}
\title{Calculate clustering thresholds for each taxon}
\usage{
superrank_thresholds(supertaxa, superrank, rank, optima)
}
\arguments{
\item{supertaxa}{(\code{character} vector) names of taxa at rank \code{superrank}}

\item{superrank}{(\code{character} string) the rank of the taxon within which the
sequences are to be clustered}

\item{rank}{(\code{character} string) the rank which is to be approximated by
clustering}

\item{optima}{(\code{data.frame}) optimum clustering thresholds to approximate
taxa at different ranks within the taxonomic hierarchy; as returned by
\code{optimize_thresholds()}}
}
\value{
(\code{numeric}) the optimal threshold for clustering sequences within
each \code{supertaxon}, in order to approximate rank \code{rank}
}
\description{
Calculate clustering thresholds for each taxon
}
\keyword{internal}
