% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_helpers.R
\name{superranks}
\alias{superranks}
\alias{subranks}
\title{Get superordinate or subordinate taxonomic ranks}
\usage{
superranks(x, ranks)

subranks(x, ranks)
}
\arguments{
\item{x}{(\code{character} string) the taxonomic rank to find superordinate or
subordinate ranks for}

\item{ranks}{(\code{character} vector) the full list of taxonomic ranks, in order
from most inclusive to least inclusive}
}
\value{
a \code{character} vector of superordinate or subordinate taxonomic ranks
}
\description{
Get superordinate or subordinate taxonomic ranks
}
\keyword{internal}
