% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_linkage.R
\name{usearch_single_linkage}
\alias{usearch_single_linkage}
\title{Do single-linkage clustering at a series of increasing similarity thresholds.}
\usage{
usearch_single_linkage(
  seq,
  seq_id = names(seq),
  method = c("tree", "matrix"),
  output_type = c("matrix", "hclust"),
  thresh_max = NULL,
  thresh_min = NULL,
  thresh_step = NULL,
  thresholds = NULL,
  precision = if (is.null(thresholds)) NULL else 0.001,
  thresh_names = names(thresholds),
  which = TRUE,
  ncpu = local_cpus(),
  usearch = Sys.which("usearch")
)
}
\arguments{
\item{seq}{(`character` vector, filename, or `Biostrings::DNAStringSet`)
sequences to cluster}

\item{seq_id}{(`character` vector) names for the sequences.  If they are
already named, this will replace the names.  Has no effect if `seq` is a
filename.}

\item{thresh_max}{(`numeric`) maximum sequence dissimilarity threshold for
clustering. Number between 0 and 1.}

\item{thresh_min}{(`numeric`) minimum sequence dissimilarity threshold for
clustering. Number between 0 and 1.}

\item{thresh_step}{(`numeric`) difference between successive percentage similarity thresholds for
clustering. Number <= thresh_max - thresh_min.}

\item{which}{(`logical`, `character` or `integer` vector, or a list of these)
subset(s) of `seq` to operate on.  If `seq` is a filename, then the distance
matrix for the full file will be calculated, but only the selected sequences
will be clustered.  Otherwise, distances will only be calculated for
sequences occurring in the subset(s)}

\item{ncpu}{(`integer` scalar) number of threads to use for calculating the
distance matrix and clustering}

\item{usearch}{(`character` scalar) path to usearch executable}

\item{thresh_name}{(`character` vector) names for the thresholds.}

\item{...}{}
}
\value{
`integer` matrix giving clustering results
}
\description{
Do single-linkage clustering at a series of increasing similarity thresholds.
}
