% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_linkage.R
\name{verify_which}
\alias{verify_which}
\title{Verify that \code{which} is a valid list of sequence subsets}
\usage{
verify_which(which, seqnames)
}
\arguments{
\item{which}{(\code{list} of \code{character} vectors) The list of sequence subsets.}

\item{seqnames}{(\code{character} vector) The names of the sequences.}
}
\value{
(\code{logical} scalar) \code{TRUE} (invisibly), or an error is thrown if
\code{which} is not a valid list of sequence subsets.
}
\description{
Verify that \code{which} is a valid list of sequence subsets
}
