% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{verify_which_impl}
\alias{verify_which_impl}
\title{Verify that \code{which} is a valid list of sequence subsets}
\usage{
verify_which_impl(which, seqnames)
}
\arguments{
\item{which}{(\code{Rcpp::List}) The list of sequence subsets.}

\item{seqnames}{(\code{Rcpp::CharacterVector}) The names of the sequences.}
}
\value{
(\code{void}) Throws an error if \code{which} is not a valid list of sequence subsets.
}
\description{
Verify that \code{which} is a valid list of sequence subsets
}
