% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{write_sequence}
\alias{write_sequence}
\title{Write sequences as a fasta file}
\usage{
write_sequence(seq, fname, ...)
}
\arguments{
\item{seq}{(\code{data.frame}, \code{character}, or \code{XStringSet}) sequences to write}

\item{fname}{(\code{character} string) file path to write to}

\item{...}{additional arguments to pass to the writing function}
}
\description{
Write sequences as a fasta file
}
