% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimotu.R
\name{closed_ref_cluster}
\alias{closed_ref_cluster}
\title{Perform closed-reference pseudo-single-linkage clustering by iterative
searching}
\usage{
closed_ref_cluster(query, ref, threshold, ...)
}
\arguments{
\item{query}{(file name, named \code{character}, \code{data.frame} or
\code{\link[Biostrings:XStringSet-class]{DNAStringSet}}) sequences to search for}

\item{ref}{(file name, named \code{character}, \code{data.frame} or
\code{\link[Biostrings:XStringSet-class]{DNAStringSet}}) sequences to search within}

\item{threshold}{(\code{numeric}) distance threshold for clustering}

\item{...}{additional arguments to pass to \code{seq_search()}}
}
\value{
(\code{data.frame}) table of sequence IDs and their cluster assignments.
}
\description{
When a sequence is equally distance from multiple references, a random
reference is chosen.
}
