% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{dist_config}
\alias{dist_config}
\alias{dist_wfa2}
\alias{dist_edlib}
\alias{dist_hybrid}
\alias{dist_hamming}
\alias{dist_usearch}
\alias{dist_file}
\title{Configuration helper for pairwise distance methods}
\usage{
dist_config(
  method = c("wfa2", "edlib", "hybrid", "hamming", "file", "usearch"),
  ...
)

dist_wfa2(
  match = 0L,
  mismatch = 1L,
  gap_open = 0L,
  gap_extend = 1L,
  gap_open2 = gap_open,
  gap_extend2 = gap_extend
)

dist_edlib()

dist_hybrid(cutoff = 0.1)

dist_hamming(min_overlap = 0L, ignore_gaps = TRUE)

dist_usearch(usearch = Sys.which("usearch"), usearch_ncpu = NULL)

dist_file(filename, by_name = TRUE)
}
\arguments{
\item{method}{(\code{character} string) method to use for distance calculations}

\item{...}{passed on to variants}

\item{match}{(non-negative \code{integer} scalar) score for a match}

\item{mismatch}{(positive \code{integer} scalar) penalty for a mismatch}

\item{gap_open}{(non-negative \code{integer} scalar) penalty for gap opening}

\item{gap_extend}{(positive \code{integer} scalar) penalty for gap extension}

\item{gap_open2}{(non-negative \code{integer} scalar) alternate penalty for gap
opening}

\item{gap_extend2}{(positive \code{integer} scalar) alternate penalty for gap
extension}

\item{cutoff}{(\code{numeric} scalar) A pairwise distance to decide when sequences
are considered "close"; when the sequences are expected to be "close", then
WFA2 in edit-distance mode is used, otherwise EdLib.}

\item{min_overlap}{(\code{integer} scalar) minimum length of overlap between two
sequences (i.e., length after subtracting end gaps in either sequence). If
the required overlap is not met, the distance is returned as 1.}

\item{ignore_gaps}{(\code{logical} flag) if \code{TRUE}, columns containing an internal
gap in one sequence are ignored, i.e. not counted as a mismatch and also not
counted towards the total aligned length. If \code{FALSE}, such columns are
counted as mismatches.  End gaps in either sequence, and internal gaps shared
by both sequences, are always ignored.}

\item{usearch}{(\code{character} path) path to the USEARCH executable}

\item{usearch_ncpu}{(\code{integer} scalar or \code{NULL}) number of threads for
USEARCH. If \code{NULL}, USEARCH will autodetect.}

\item{filename}{(\code{character} string) path to the distance matrix file}

\item{by_name}{(\code{logical} flag) if \code{TRUE}, the sequences are identified in
the file by name; if \code{FALSE}, they are identified by number, starting with 0.
Identification by number may be slightly faster and reduces the necessary
file size, but is not supported by \code{optimotu()}, only by
\code{optimize_thresholds()}.}
}
\value{
an object representing the pairwise distance method
}
\description{
Configuration helper for pairwise distance methods
}
\section{Functions}{
\itemize{
\item \code{dist_wfa2()}: helper function for method \code{"wfa2"}

\item \code{dist_edlib()}: helper function for method \code{"edlib"}

\item \code{dist_hybrid()}: helper function for method \code{"hybrid"}

\item \code{dist_hamming()}: helper function for method \code{"hamming"}

\item \code{dist_usearch()}: helper function for method \code{"usearch"}

\item \code{dist_file()}: helper function for method \code{"file"}

}}
\references{
Edgar, R.C., 2010. Search and clustering orders of magnitude faster than
BLAST. Bioinformatics 26, 2460–2461.
https://doi.org/10.1093/bioinformatics/btq461

Marco-Sola, S., Moure, J.C., Moreto, M., Espinosa, A., 2021. Fast gap-affine
pairwise alignment using the wavefront algorithm. Bioinformatics 37,
456–463. https://doi.org/10.1093/bioinformatics/btaa777

Šošić, M., Šikić, M., 2017. Edlib: a C/C ++ library for fast, exact sequence
alignment using edit distance. Bioinformatics 33, 1394–1395.
https://doi.org/10.1093/bioinformatics/btw753
}
