% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_thresholds.R
\name{optimize_thresholds}
\alias{optimize_thresholds}
\title{Optimize clustering thresholds for taxonomically identified reference sequences}
\usage{
optimize_thresholds(
  taxonomy,
  refseq,
  ranks = c("kingdom", "phylum", "class", "order", "family", "genus", "species"),
  dist_config = dist_wfa2(),
  threshold_config = threshold_uniform(0, 0.4, 0.001),
  clust_config = clust_tree(),
  parallel_config = parallel_concurrent(threads = 1),
  min_taxa = 5L,
  min_refseq = 2L * min_taxa,
  id_col = "seq_id",
  verbose = FALSE
)
}
\arguments{
\item{taxonomy}{(\code{data.frame}) taxonomic identifications of the reference
sequences.  Must include a column for each rank in \code{ranks} and a column for
sequence identifiers, defined by \code{id_col}. Any additional columns are ignored.}

\item{refseq}{(named \code{character}) the reference sequences}

\item{ranks}{(\code{character} vector) the taxonomic ranks in \code{taxonomy}}

\item{dist_config}{(\code{optimotu_dist_config}) specification of the pairwise
distance algorithm to use, as created by \code{dist_config()} or its helpers}

\item{threshold_config}{(\code{optimotu_threshold_config}) specification of the
thresholds to test, as created by \code{threshold_config()} or its helpers}

\item{clust_config}{(\code{optimotu_clust_config}) specification of the clustering
algorithm to use, as created by \code{clust_config()} or its helpers}

\item{parallel_config}{(\code{optimotu_parallel_config}) specification of the
parallelization scheme to use, as created by \code{parallel_config()} or its
helpers}

\item{min_taxa}{(\code{integer(1)}) the minimum number of subtaxa at a given rank
which must belong to a taxon to optimize thresholds for that rank within that
taxon. Must be at least 2.}

\item{min_refseq}{(\code{integer(1)}) the minimum number of reference sequences
which must belong to a taxon to optimize thresholds for that taxon. Must be
at least \code{min_taxa} (but should probably be more).}

\item{id_col}{(\code{character(1)}) the name of the column in \code{taxonomy} which
contains sequence identifiers}

\item{verbose}{(\code{logical(1)} or \code{integer(1)}) whether to print progress
messages; values greater than 1 (or TRUE) print more}
}
\value{
(\code{data.frame}) a data frame with the following columns:
\itemize{
\item \code{rank} (\code{character}) the rank being optimized
\item \code{superrank} (\code{character}) the containing rank for optimization
\item \code{supertaxon} (\code{character}) the containing taxon
\item \code{metric} (\code{character}) the clustering quality metric
\item \code{threshold} (\code{numeric}) the threshold which maximizes the metric
\item \code{value} (\code{numeric}) the value of the metric at the threshold
}
}
\description{
Optimize clustering thresholds for taxonomically identified reference sequences
}
