% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{seq_as_char}
\alias{seq_as_char}
\title{get sequences as a character vector}
\usage{
seq_as_char(seq, seq_col = NULL, name_col = NULL)
}
\arguments{
\item{seq}{(\code{data.frame}, \code{character}, or \code{XStringSet}) sequences to get}

\item{seq_col}{(\code{character}) name of the column in a data frame which
contains the sequences}

\item{name_col}{(\code{character}) name of the column in a data frame which
contains the sequence names}
}
\description{
get sequences as a character vector
}
