% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_cluster.R
\name{seq_cluster}
\alias{seq_cluster}
\title{Single-linkage clustering of nucleotide sequences}
\usage{
seq_cluster(
  seq,
  seq_id = names(seq),
  dist_config,
  threshold_config,
  clust_config = clust_tree(),
  parallel_config = parallel_concurrent(1),
  output_type = c("matrix", "hclust"),
  which = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{seq}{(\code{character} vector, filename,
\link[Biostrings:XStringSet-class]{DNAStringSet}, or \code{data.frame} with columns
"seq_id" (\code{character}) and "seq" (\code{character})) sequences to cluster}

\item{seq_id}{(\code{character} vector) names for the sequences.  If they are
already named, this will replace the names.  Has no effect if \code{seq} is a
filename.}

\item{dist_config}{(\code{optimotu_dist_config} object returned by
\code{\link[=dist_config]{dist_config()}} or one of its helpers) Configuration of the method to
calculate distances.}

\item{threshold_config}{(\code{optimotu_threshold_config} object returned by
\code{\link[=threshold_config]{threshold_config()}} or one of its helper functions) Definition of the
thresholds to use for clustering.}

\item{clust_config}{(\code{optimotu_cluster_config} object returned by
\code{\link[=clust_config]{clust_config()}} or one of its helpers) The clustering algorithm to use; all
algorithms give identical results, but may have different performance
characteristics on different problems.}

\item{parallel_config}{(\code{optimotu_parallel_config} object returned by
\code{\link[=parallel_config]{parallel_config()}} or one of its helpers) The method to use for
parallel clustering. For single-threaded clustering, use the default value
(\code{\link[=parallel_concurrent]{parallel_concurrent()}} with \code{threads = 1}).}

\item{output_type}{(\code{character}) Which type of output to give; one of
\code{"matrix"} or \code{"hclust"}. \code{"matrix"} returns an integer matrix giving
clustering results, where the element in row \code{i} and column \code{j} gives the
0-based index of the first member of the cluster to which sequence \code{j}
belongs when clustered at the \code{i}th clustering threshold. \code{"hclust"} returns
an object as returned by \code{\link[stats:hclust]{stats::hclust()}}, which requires less memory,
especially for large problems. If \code{which} is given, then both \code{output_type}s
instead return a list whose elements are of the chosen type.}

\item{which}{(\code{list} of \code{character} vectors) Instead of performing clustering
on all input sequences, perform independent clustering on subsets of the
sequences defined by the elements of \code{which}. Subsets do not need to be
disjoint (and indeed, if they are it is probably faster to calculate the
distance matrices separately.)}

\item{verbose}{(\code{logical} flag)}
}
\description{
Single-linkage clustering of nucleotide sequences
}
