% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/seq_distmx.R
\name{seq_distmx}
\alias{seq_distmx}
\alias{seq_distmx_edlib}
\alias{seq_distmx_kmer}
\alias{seq_distmx_wfa2}
\alias{seq_distmx_hamming}
\title{Sparse distance matrix between DNA sequences}
\usage{
seq_distmx_edlib(
  seq,
  dist_threshold,
  details = 0L,
  span = 0L,
  constrain = TRUE,
  threads = 1L,
  verbose = 0L
)

seq_distmx_kmer(
  seq,
  dist_threshold,
  udist_threshold,
  match = -1L,
  mismatch = 2L,
  gap_open = 10L,
  gap_extend = 1L,
  gap_open2 = 0L,
  gap_extend2 = 0L,
  threads = 1L
)

seq_distmx_wfa2(
  seq,
  dist_threshold,
  match = -1L,
  mismatch = 2L,
  gap_open = 10L,
  gap_extend = 1L,
  gap_open2 = 0L,
  gap_extend2 = 0L,
  prealign = TRUE,
  constrain = TRUE,
  threads = 1L
)

seq_distmx_hamming(
  seq,
  dist_threshold,
  min_overlap = 0L,
  ignore_gap = FALSE,
  threads = 1L,
  verbose = 0L
)

seq_distmx(
  seq,
  threshold,
  seq_id = NULL,
  dist_config = dist_wfa2(),
  parallel_config = parallel_concurrent(1),
  verbose = FALSE,
  details = c("none", "gapstats", "cigar"),
  span = c("global", "extension"),
  constrain = TRUE,
  id_is_integer = is.data.frame(seq) && "seq_idx" \%in\% names(seq),
  ...
)
}
\arguments{
\item{seq}{(\code{character}, \code{data.frame}, or \code{XStringSet}) sequences to search}

\item{dist_threshold}{(\code{numeric} scalar) maximum sequence distance (edit
distance / alignment length) threshold for reporting}

\item{details}{(one of "none", "gapstats", or "cigar") if "gapstats", return
#' the gap statistics. If "cigar", return the CIGAR string. Otherwise return
only the score and distance}

\item{span}{(\code{character} string) the span of the alignment; currently
accepted values are "global" and "extension".  The default is "global".}

\item{constrain}{(\code{logical} flag) if \code{TRUE}, the alignment algorithm will
use optimizations that will cause it to exit early if the optimal alignment
has a distance greater than the distance threshold. This should not change
the correctness of distance calculations below the threshold, and results in
a large speedup. It is recommended to use \code{constrain=FALSE} only to verify
that the results do not change.}

\item{threads}{(\code{integer} count) number of parallel threads to use for
computation.}

\item{verbose}{(\code{logical} or \code{integer} scalar) print progress messages.}

\item{udist_threshold}{(\code{numeric} scalar between 0 and 1) maximum udist
(number of shared kmers / number of kmers in the shorter sequence) for full
alignment.}

\item{match}{(\code{integer} scalar) score for a match, default: 1}

\item{mismatch}{(\code{integer} scalar) score for a mismatch, default: 2}

\item{gap_open}{(\code{integer} scalar) score for opening a gap, default: 10}

\item{gap_extend}{(\code{integer} scalar) score for extending a gap, default: 1}

\item{gap_open2}{(\code{integer} scalar) alternate score for opening a gap,
default: 0}

\item{gap_extend2}{(\code{integer} scalar) alternate score for extending a gap,
default: 0}

\item{prealign}{(\code{logical} flag) if \code{TRUE}, do a prealignment using
edit-distance as alignment score (\verb{match = 0, mismatch = 1, gap_open = 0, gap_extend = 1, gap_open2 = 0, gap_extend2 = 1}) to test feasibility before
aligning with alternate alignment scores. Note that, since pairwise distance
is defined using edit distance, any other set of scores will always result
in a pairwise distance which is equal to or greater than an alignment based
on the edit distance score.}

\item{threshold}{(\code{numeric} scalar) maximum distance to consider a match, in
[0, 1] where 0 is identical.}

\item{seq_id}{(\code{character} vector) names for the sequences.  If they
are already named, this will replace the names.}

\item{dist_config}{(\code{optimotu_dist_config}) configuration for calculating
distances, as returned by \code{dist_config()} or its helpers.}

\item{parallel_config}{(\code{optimotu_parallel_config}) configuration for
parallel processing, as returned by \code{parallel_config()} or its helpers.}

\item{id_is_integer}{(\code{logical} scalar) if \code{TRUE}, the sequence IDs are
parsed as integers, and the returned IDs are integers.
The default is \code{FALSE}.}

\item{...}{passed to methods}
}
\value{
(\code{data.frame}) a sparse distance matrix; columns are \code{seq1} and
\code{seq2} for the 0-based indices of two sequences; \code{score1} and \code{score2} are
the optimal alignment score for the two sequences in the "prealignment" (if
any) and "alignment" stages; \code{dist1} and \code{dist2} are the corresponding
sequence distances.

(\code{data.frame}) with columns "seq_id1" and "seq_id2" (\code{character}),
or "seq_idx1" and "seq_idx2" (\code{integer}) if \code{id_is_integer} is \code{TRUE},
"score1" and "dist1" giving the score and distance for the prealignment
stage, if any, and "score2" and "dist2" (\code{numeric}) giving the final score
and distance. If \code{details} is "gapstats", the gap statistics are returned in
columns "align_length", "n_insert", "n_delete", "max_insert", and
"max_delete" (\code{integer}). If \code{details} is "cigar", the CIGAR string is
returned in a column "cigar" (\code{character}).
}
\description{
Sparse distance matrix between DNA sequences

Search for the closest match(es) to sequences in a reference database
}
