% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_search.R
\name{seq_search}
\alias{seq_search}
\title{Search for the closest match(es) to sequences in a reference database}
\usage{
seq_search(
  query,
  ref,
  threshold,
  query_id = NULL,
  ref_id = NULL,
  dist_config = dist_wfa2(),
  parallel_config = parallel_concurrent(1),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{query}{(\code{character}, \code{data.frame}, or \code{XStringSet}) sequences to search}

\item{ref}{(\code{character}, \code{data.frame}, or \code{XStringSet}) reference sequences to
search against}

\item{threshold}{(\code{numeric} scalar) maximum distance to consider a match, in
[0, 1] where 0 is identical.}

\item{query_id}{(\code{character} vector) names for the query sequences.  If they
are already named, this will replace the names.}

\item{ref_id}{(\code{character} vector) names for the reference sequences.  If they
are already named, this will replace the names.}

\item{dist_config}{(\code{optimotu_dist_config}) configuration for calculating
distances, as returned by \code{dist_config()} or its helpers.}

\item{parallel_config}{(\code{optimotu_parallel_config}) configuration for parallel
processing, as returned by \code{parallel_config()} or its helpers.}

\item{verbose}{(\code{logical} or \code{integer} scalar) print progress messages.}

\item{...}{passed to methods}
}
\value{
(\code{data.frame}) with columns "seq_id" (\code{character}),
"ref_id" (\code{character}), and "dist" (\code{numeric}) giving the distance between
the query and reference.
}
\description{
Search for the closest match(es) to sequences in a reference database
}
