% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{summarize_by_rank}
\alias{summarize_by_rank}
\title{Summarize taxonomic ranks by superordinate rank}
\usage{
summarize_by_rank(data, ranks)
}
\arguments{
\item{data}{('data.frame') the taxonomy to summarize; should contain a
column named \code{seq_id} and columns for each value of \code{ranks}}

\item{ranks}{('character') the ranks to summarize}
}
\value{
a data frame with columns:
\itemize{
\item \code{supertaxon} (\code{character}) the superordinate taxon
\item \code{superrank} (\code{character}) the rank of the superordinate taxon
\item \code{rank} (\code{character}) the rank being summarized
\item \code{n_taxa} (\code{integer}) the number of unique taxa at the rank
\item \code{n_seq} (\code{integer}) the number of sequences
\item \code{seq_id} (\code{list} of \code{character}) a list of sequence IDs
\item \code{true_parition} (\code{list} of \code{integer}) integer mapping to taxa for each
element in \code{seq_id}
}
}
\description{
Summarize taxonomic ranks by superordinate rank
}
\keyword{internal}
