% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_linkage.R
\name{usearch_single_linkage}
\alias{usearch_single_linkage}
\title{Do single-linkage clustering at a series of increasing similarity thresholds
using USEARCH to calculate sequence similarities.}
\usage{
usearch_single_linkage(
  seq,
  seq_id = names(seq),
  method = c("tree", "matrix"),
  output_type = c("matrix", "hclust"),
  thresh_max = NULL,
  thresh_min = NULL,
  thresh_step = NULL,
  thresholds = NULL,
  precision = if (is.null(thresholds)) NULL else 0.001,
  thresh_names = names(thresholds),
  which = TRUE,
  ncpu = local_cpus(),
  usearch = Sys.which("usearch")
)
}
\arguments{
\item{seq}{(\code{character} vector, filename,
\link[Biostrings:XStringSet-class]{DNAStringSet}, or \code{data.frame} with columns
"seq_id" (\code{character}) and "seq" (\code{character})) sequences to cluster}

\item{seq_id}{(\code{character} vector) names for the sequences.  If they are
already named, this will replace the names.  Has no effect if \code{seq} is a
filename.}

\item{thresh_max}{(\code{numeric}) maximum sequence dissimilarity threshold for
clustering. Number between 0 and 1.}

\item{thresh_min}{(\code{numeric}) minimum sequence dissimilarity threshold for
clustering. Number between 0 and 1.}

\item{thresh_step}{(\code{numeric}) difference between successive percentage similarity thresholds for
clustering. Number <= thresh_max - thresh_min.}

\item{which}{(\code{logical}, \code{character} or \code{integer} vector, or a list of these)
subset(s) of \code{seq} to operate on.  If \code{seq} is a filename, then the distance
matrix for the full file will be calculated, but only the selected sequences
will be clustered.  Otherwise, distances will only be calculated for
sequences occurring in the subset(s)}

\item{ncpu}{(\code{integer} scalar) number of threads to use for calculating the
distance matrix and clustering}

\item{usearch}{(\code{character} scalar) path to usearch executable}

\item{thresh_name}{(\code{character} vector) names for the thresholds.}
}
\value{
\code{integer} matrix giving clustering results
}
\description{
This function uses a unix pipe to direct the output of the
USEARCH "\code{calc_distmx}" command  to \code{\link[=single_linkage]{single_linkage()}}. USEARCH (version 8.0
or higher) should be installed separately; it is available with a free
license for most users at \link{https://www.drive5.com/usearch/}.
}
